feat: implement prioritized streaming mode for media playback
