+++ aura-core/src/config/bittorrent.rs
@@ -33,6 +33,7 @@ pub struct BitTorrentConfig {
     pub endgame_mode_enabled: bool,
     pub endgame_threshold_pieces: usize,
     pub endgame_threshold_percent: f32,
+    pub streaming_metadata_pieces: usize,
     pub min_split_size_mb: u64,
     pub max_connections_per_torrent: usize,
     pub peer_id_prefix: String,
@@ -63,6 +64,7 @@ impl Default for BitTorrentConfig {
             endgame_mode_enabled: true,
             endgame_threshold_pieces: 3,
             endgame_threshold_percent: 0.01,
+            streaming_metadata_pieces: 4,
             min_split_size_mb: 20,
             max_connections_per_torrent: 200,
             peer_id_prefix: "-AR0001-".to_string(),
