+++ aura-core/src/worker/bittorrent/task/mod.rs
@@ -2,8 +2,10 @@ pub mod choker;
 pub mod dht;
 pub mod logic;
 pub mod lpd;
+pub mod state;
 pub mod tracker;
 pub use logic::*;
+pub use state::*;
 
 #[cfg(test)]
 mod tests;
